third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-invalid-integrity-expected.txt

CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256:yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8='). The hash algorithm must be followed by a ';' character.
CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256;(yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8=)'). The digest must be a valid, base64-encoded value.
This test passes if no 'FAIL' alert appears.
